move svg.tsx and change it's imports
